==> doc graph/dataset_builder/0_README.ipynb <==
# Readme 

This is a package of ipython notebooks intended to recreate the neo4j database for the DocGraph API (located [here](https://github.com/Analytics-Innovation-Incubator/Cohort-Builder)). There are datasets and python packages required to run these notebooks. See below...

## Instructructions
1. Download datasets and extract appropriate files to the _'data'_ folder
2. Install Python packages required _(if necessary)_
3. Install neo4j _(if necessary)_
4. Execute each ipython notebook in numericalorder (1-5)
5. Compile graph.db from files generated (see notebook #6 for instructions)

## Requirements

### Datasets Required
</p>
1. Download each dataset below (they're on PwC google drive, so use that authentication)
2. Extract each of the zip files
3. Take the file with the name indicated in the sub-bullet and save it to the <i>'data'</i> folder in this directory
</p>

- <b>DocGraph Edge Data</b> - https://drive.google.com/open?id=0B15A78kQcodoaDFGWUJZc2ZjdDQ&authuser=0
    - <i>DocGraph-2012-2013-Days365.csv</i>
- <b>DocGraph Node Data</b> - https://drive.google.com/open?id=0B15A78kQcododGU4MlVSTmtLdzQ&authuser=0
    - <i>DocGraph_Procedure.csv</i>
- <b>NPPES Data</b> - https://drive.google.com/open?id=0B5VudqrtA4nUdDduQ3hnSVhUTnM&authuser=0
    - <i>npidata_20050523-20150208.csv</i>
- <b>HCPCS Description Data</b> - https://drive.google.com/open?id=0B5VudqrtA4nUb01xUXNZcmpva2c&authuser=0
    - <i>Medicare-Physician-and-Other-Supplier-PUF-CY2012.txt</i>

### Python Packages Required

- Graphlab-create 1.3 or higher
    - https://github.com/dato-code/Dato-Core
- pyspark 1.3 or higher (Apache-Spark)
    - https://spark.apache.org/

### neo4j Installation Required
- Install neo4j 2.2.0 or higher (community edition)
    - http://neo4j.com/download/

## Additional Notes

- These notebooks use graphlab to convert Spark RDDs to graphlab SFrames. To enable this capability, you must change your <b>Spark</b> configuration file to point Spark to a graphlab JAR file. See the <i>'Spark RDD -> Graphlab Tutorial'</i> notebook in this directory for an example of how to set it up.
- When using Spark in these notebooks, feel free to change the configuration settings when initializing the SparkContext. <b>All of the default settings are 10 GB of memory and 1 core.</b> Example below...
    - <code>conf.set("spark.executor.memory", "10g")
    - <code>conf.set("spark.cores.max", "1")
