==> doc graph/dataset_builder/1_Chicago Trim.ipynb <==
# Trimming DG to just Chicago Providers

### Imports

from pyspark import SparkContext, SparkConf
import graphlab as gl

### Setting Up Spark Instance

# Configure Spark Settings
conf=SparkConf()
conf.set("spark.executor.memory", "10g")
conf.set("spark.cores.max", "1")
conf.setAppName("My App")

## Initialize SparkContext
sc = SparkContext('local[*]', conf=conf)

### Helper Functions

# function to type cast variables
def safe_cast(val, to_type, default=0):
    try:
        return to_type(val)
    except ValueError:
        return default

# function to filter out all doctors not in Chicago
def chicago_filter(line):
    five_zip = safe_cast(line[10][0:5].replace('\"',''), int)
    if five_zip in zip_list:
        return True
    else:
        return False

# cleans unwanted quotes around data
def clean(line):
    new_line = []
    for cell in line:
        new_line.append(cell.replace('\"',''))
    return new_line

# decodes unicode
def encode_filter(line):
    new_line = []
    for first in line:
        first = first.split(',')
        for cell in first:
            cell = cell.encode('utf-8').strip()
            new_line.append(cell)    
    return new_line

# function to filter out the quotes in every line
def quote_filter(line):
    new_line = []
    for cell in line:
        new_line.append(cell.replace('\"',''))
    return new_line

# filters NPIs in both columns of the edge data
def filter_npis(line):
    if line[0] in npi_dict and line[1] in npi_dict:
        return True
    else:
        return False

## Trimming to Chicago Zip Codes

### Loading the data

# DocGraph Node Data
dg_node_raw = sc.textFile('data/DocGraph_Procedure.csv').map(lambda x: x.split('\t'))

# DocGraph Edge Data
dg_edge_raw = sc.textFile('data/DocGraph-2012-2013-Days365.csv').map(lambda x: x.split('\t'))

# Chicago Zip Codes
chicago_zips = gl.SFrame.read_csv('data/chicago_metro_zip.csv',verbose = False)
zip_list = list(chicago_zips['zip'])

### Filter the data

# filter out the header
dg_header = dg_node_raw.first()
dg_node_raw = dg_node_raw.filter(lambda line: line != dg_header)

# line by line filtering out the quotes in each cell
dg_node_updated = dg_node_raw.map(lambda x : quote_filter(x))

# filter out the zip codes not in Chicago
dg_node_chicago = dg_node_updated.filter(lambda line: chicago_filter(line))

### Convert the RDD to GraphLab SFrame

# converts an RDD to an SFrame in GL
dg_node_sf = gl.SFrame.from_rdd(dg_node_chicago)

# the data is interpreted as one column
# we can use the unpack function in GL to create multiple columns
dg_node_sf_u = dg_node_sf.unpack('X1')

# create the column name dictionary
col_names = {'X1.0':'npi','X1.1':'nppes_provider_last_org_name','X1.2':'nppes_provider_first_name','X1.3':'nppes_provider_mi',
             'X1.4':'nppes_credentials','X1.5':'nppes_provider_gender','X1.6':'nppes_entity_code','X1.7':'nppes_provider_street1',
             'X1.8':'nppes_provider_street2','X1.9':'nppes_provider_city','X1.10':'nppes_provider_zip',
             'X1.11':'nppes_provider_state','X1.12':'nppes_provider_country','X1.13':'provider_type',
             'X1.14':'medicare_participation_indicator','X1.15':'place_of_Service','X1.16':'hcpcs_code','X1.17':'line_srvc_cnt',
             'X1.18':'bene_unique_cnt','X1.19':'bene_day_srvc_cnt','X1.20':'average_Medicare_allowed_amt',
             'X1.21':'stdev_Medicare_allowed_amt','X1.22':'average_submitted_chrg_amt','X1.23':'stdev_submitted_chrg_amt',
             'X1.24':'average_Medicare_payment_amt','X1.25':'stdev_Medicare_payment_amt'}

# rename the columns
dg_node_sf_u = dg_node_sf_u.rename(col_names)

# create the column data type dictionary
col_types = {'line_srvc_cnt':int,'bene_unique_cnt':int,'bene_day_srvc_cnt':int,'average_Medicare_allowed_amt':float,
             'stdev_Medicare_allowed_amt':float,'average_submitted_chrg_amt':float,'stdev_submitted_chrg_amt':float,
             'average_Medicare_payment_amt':float,'stdev_Medicare_payment_amt':float}

# loop through the columns and give each the appropriate data type
for name in dg_node_sf_u.column_names():
    if name in col_types.keys():
        print name, 'starting...'
        if col_types[name] is int:
            dg_node_sf_u[name] = dg_node_sf_u[name].astype(int)
        elif col_types[name] is float:
            dg_node_sf_u[name] = dg_node_sf_u[name].astype(float)
        else:
            print 'Something went wrong with column:', name
        print name, 'ending...'

# create a list of unique NPIs in Chicago
chi_npi_list = list(dg_node_sf_u['npi'].unique())

# create a dictionary of these unique values for python processing speed improvements
npi_dict = {}
for npi in chi_npi_list:
    npi_dict[npi] = 0

### Transform and filter the edge data

# run each row through the encoding function
dg_edge_updated = dg_edge_raw.map(lambda x : encode_filter(x))

# run earch row through the NPI filter
dg_edge_chicago = dg_edge_updated.filter(lambda line : filter_npis(line))

### Convert the edge RDD to a GL SFrame

# convert the RDD to an SFrame using GL
dg_edge_sf = gl.SFrame.from_rdd(dg_edge_chicago)

# the data is interpreted as one column
# we can use the unpack function in GL to create multiple columns 
dg_edge_sf_u = dg_edge_sf.unpack('X1')

# create the column names dictionary and rename each column
col_names = {'X1.0':'FirstNPI','X1.1':'SecondNPI','X1.2':'SharedTransactionCount','X1.3':'PatientTotal','X1.4':'SameDayTotal'}
dg_edge_sf_u = dg_edge_sf_u.rename(col_names)

# create the column data types dictionary
col_types = {'FirstNPI':str,'SecondNPI':str,'SharedTransactionCount':int,'PatientTotal':int,'SameDayTotal':int}

# loop through each column and assign it the appropriate data type
for name in dg_edge_sf_u.column_names():
    if name in col_types.keys():
        print name, 'starting...'
        if col_types[name] is int:
            dg_edge_sf_u[name] = dg_edge_sf_u[name].astype(int)
        elif col_types[name] is float:
            dg_edge_sf_u[name] = dg_edge_sf_u[name].astype(float)
        elif col_types[name] is str:
            dg_edge_sf_u[name] = dg_edge_sf_u[name].astype(str)
        else:
            print 'Something went wrong with column:', name
        print name, 'ending...'

### Save the resulting data

# save the filtered nodes and edges
dg_node_sf_u.save('data/dg_nodes_chicago.csv')
dg_edge_sf_u.save('data/dg_edges_chicago.csv')

# save the unique list of NPIs
dg_node_sf_u[['npi']].unique().save('data/unique_npis.csv')
