==> doc graph/dataset_builder/2_Filter, Clean, Join NPPES Data.ipynb <==
# Filter, Clean, and Join the NPPES Data w/ DocGraph Data

### Imports

import graphlab as gl
from pyspark import SparkConf,SparkContext
import os

### Initialize Spark Instance

# Configure Spark Settings
conf=SparkConf()
conf.set("spark.executor.memory", "10g")
conf.set("spark.cores.max", "1")
conf.setAppName("My App")

## Initialize SparkContext
sc = SparkContext('local[*]', conf=conf)

### Helper Functions

# function to type cast variables
def safe_cast(val, to_type, default=0):
    try:
        return to_type(val)
    except ValueError:
        return default

# cleans unwanted quotes around data
def clean(line):
    new_line=[]
    for i in line:
        i=i.replace('\"','')
        new_line.append(i)
    return new_line

# filters the NPIs to the ones we care about
def filter_npis(line):
    npi = safe_cast(clean(line)[0],int)
    rep_npi = safe_cast(clean(line)[2],int)
    
    if (npi in npi_dict) or (rep_npi in npi_dict):
        return True
    else:
        return False

### Do some preliminary cleaning to the NPPES data

# open the NPPES data
with open('data/npidata_20050523-20150208.csv','r') as f:
    nppes_pre = f.readlines()

# save the data back to a new CSV file as tab delimited
with open('data/nppes_fixed.csv','w') as f:
    for line in nppes_pre:
        f.write(line.replace('\","',"\t").replace('\"',''))

### Ingest the pre-cleaned data

# read the CSV file into Spark
raw_data = sc.textFile('data/nppes_fixed.csv').map(lambda x: x.split('\t'))

# get the header for the file
header = raw_data.first()

# get the unique NPI list from our data
npi_list = gl.SFrame.read_csv('data/unique_npis.csv', verbose = False)

# put the unique NPI list into a dictionary for quicker python processing speeds
npi_dict = {}
for line in npi_list['npi']:
    npi = safe_cast(line,int)
    npi_dict[npi] = 0

# filter the data by only the NPIs in Chicago
filtered_data = raw_data.filter(lambda line: filter_npis(line))

### Convert the RDD to an SFrame in GL

# convert rdd to SFrame
nppes_sf = gl.SFrame.from_rdd(filtered_data)

# the data is interpreted as one column
# we can use the unpack function in GL to create multiple columns
nppes_sf_u = nppes_sf.unpack('X1')

# create a header dictionary
header_dict = {}
y = 0
for x in header:
    header_dict[y] = x
    y += 1

# use the header dictionary to rename the columns
for col in nppes_sf_u.column_names():
    nppes_sf_u.rename({col:header_dict[int(col[3:])]})

# remove columns that have zero data
for col in nppes_sf_u.column_names():
    counter = 0
    temp_data = nppes_sf_u[col]
    for cell in temp_data:
        if cell != "":
            counter += 1
        if counter > 0:
            break
    
    if counter == 0:
        nppes_sf_u.remove_column(col)
        print "Removing column:", col
    continue

# rename the joining column name to match the DG dataset
nppes_sf_u = nppes_sf_u.rename({'NPI':'npi'})

### Read in the DG Chicago Nodes

# read in the docgraph data
dg_chicago_nodes = gl.SFrame.read_csv('data/dg_nodes_chicago.csv', verbose = False)

### Join the data on the NPI

# use join to combine the two files via the unique identifier (NPI)
joined_data = dg_chicago_nodes.join(nppes_sf_u,on = 'npi')

### Save the resulting data

# save the resulting dataset
joined_data.save('data/dg_nppes_chicago_joined.csv')
