==> doc graph/dataset_builder/3_Decode SK&A Data.ipynb <==
# Decode the SK&A Data

### Imports

import graphlab as gl

## Load in the SK&A Data

# load the SK&A data into GL
ska_data_raw = gl.SFrame.read_csv('data/Provider Extract - Chicago NPIs.csv', verbose=False)

# rename the unique ID to what we're used to
ska_data_raw = ska_data_raw.rename({'NPI':'npi'})

## Decode SK&A Data

#### The code does the same thing for each column needing decoding. Therefore I didn't comment the code below
- brings in the data
- renames the columns to something more meaningful
- sometimes gets rid of blank values
- joins those columns to the larger data

### CODE 2

c2 = gl.SFrame.read_csv('data/SK&A Codes/CODE2.csv',verbose = False)

c2 = c2.rename({'Code2':'CODE2','Practice Specialty':'Practice Specialty Value'})

ska_data_c2 = ska_data_raw.join(c2,'CODE2').rename({'CODE2':'Practice Specialty Code'})

### CODE 3

c3 = gl.SFrame.read_csv('data/SK&A Codes/CODE3.csv', verbose = False)

c3 = c3.rename({'EXPL3':'PGP Value','STATE':'PGP State'}).remove_column('X4')

c3 = c3.filter_by('','CODE3',exclude = True)

c3 = c3.append(gl.SFrame({'CODE3':[""],'PGP Value':[""],'PGP State':[""]}))

ska_data_c2c3 = ska_data_c2.join(c3,'CODE3','left').rename({'CODE3':'PGP Code'})

#### CODE 4

c4 = gl.SFrame.read_csv('data/SK&A Codes/CODE4.csv', verbose = False)

c4 = c4.rename({'Code4':'CODE4','STATE':'Health System State','Health System Code':'Health System Value'})

c4 = c4.filter_by('','CODE4',exclude = True)

c4 = c4.append(gl.SFrame({'CODE4':[""],'Health System Value':[""],'Health System State':[""]}))

ska_data_c2c3c4 = ska_data_c2c3.join(c4,'CODE4','left').rename({'CODE4':'Health System Code'})

## Code 5

c5 = gl.SFrame.read_csv('data/SK&A Codes/CODE5.csv', verbose = False)

c5 = c5.rename({"Code5":"CODE5",'Hospital Code':'Hospital Value','STATE':'Hospital State'})

c5 = c5.append(gl.SFrame({'CODE5':[""],'Hospital Value':[""],'Hospital State':[""]}))

ska_data_c2c3c4c5 = ska_data_c2c3c4.join(c5,'CODE5','left').rename({'CODE5':'Hospital Code'})

## Code 6

c6 = gl.SFrame.read_csv('data/SK&A Codes/CODE6.csv', verbose = False)

c6 = c6.rename({'Code6':'CODE6','IPA Code':'IPA Value','STATE':'IPA State'})

c6 = c6.append(gl.SFrame({'CODE6':[""],'IPA Value':[""],'IPA State':[""]}))

ska_data_c2c3c4c5c6 = ska_data_c2c3c4c5.join(c6,'CODE6','left').rename({'CODE6':'IPA Code'})

## Code 7

c7 = gl.SFrame.read_csv('data/SK&A Codes/CODE7.csv', verbose = False)

c7 = c7.rename({'Code7':'CODE7','Affiliation Code':'Affiliation Value','STATE':'Affiliation State'})

c7 = c7.append(gl.SFrame({'CODE7':[""],'Affiliation Value':[""],'Affiliation State':[""]}))

ska_data_c2c3c4c5c6c7 = ska_data_c2c3c4c5c6.join(c7,'CODE7','left').rename({'CODE7':'Affiliation Code'})

## SPEC

spec = gl.SFrame.read_csv('data/SK&A Codes/SPEC.csv', verbose = False)

spec = spec.rename({'Spec':'SPEC','Physician Specialty':'Physician Specialty Value'})

ska_data_c2c3c4c5c6c7S = ska_data_c2c3c4c5c6c7.join(spec,'SPEC','left').rename({'SPEC':'Physician Specialty Code'})

### Save the final output

ska_data_c2c3c4c5c6c7S.save('data/SK&A Data Decoded.csv')
