==> doc graph/dataset_builder/4_HCPCS Description File Creator.ipynb <==
# HCPCS Description Grabber 

### Imports

import graphlab as gl

### Data Ingest

# grab the file with all the HCPCS descriptions
big_data = gl.SFrame.read_csv('data/Medicare-Physician-and-Other-Supplier-PUF-CY2012.txt',
                              delimiter = '\t', verbose = False)

### Data Filter

# select the two columns we care about and exclude any blanks
hcpcs = big_data.select_columns(['hcpcs_code','hcpcs_description']).filter_by('','hcpcs_code',exclude=True)

# get the unique pairs
hcpcs_u = hcpcs.unique()

# sort the data by the HCPCS code
final = hcpcs_u.sort('hcpcs_code')

### Save the resulting dataset

# save the output
final.save('data/hcpcs_descriptions.csv')
