==> doc graph/dataset_builder/5_Create Nodes & Edges for Database Creation.ipynb <==
# Create the Nodes & Rels Files for neo4j Import 

### Imports

import graphlab as gl
from __future__ import division

# DocGraph Nodes & Rels

### Import the data

# import the joined dg data
docgraph_node_raw = gl.SFrame.read_csv('data/dg_nppes_chicago_joined.csv', delimiter=',',
                                        column_type_hints={'line_srvc_cnt':int, 
                                                          'bene_unique_cnt':int, 
                                                          'bene_day_srvc_cnt':int,
                                                          'average_Medicare_allowed_amt':float,
                                                          'stdev_Medicare_allowed_amt':float,
                                                          'average_submitted_chrg_amt':float,
                                                          'stdev_submitted_chrg_amt':float,
                                                          'average_Medicare_payment_amt':float,
                                                          'stdev_Medicare_payment_amt':float},verbose=False)

# select columns wanted
df = docgraph_node_raw.select_columns(['npi','nppes_provider_last_org_name','nppes_provider_first_name',
                                       'nppes_provider_mi','nppes_credentials','nppes_provider_gender',
                                       'nppes_entity_code','nppes_provider_street1','nppes_provider_street2',
                                       'nppes_provider_city','nppes_provider_zip','nppes_provider_state',
                                       'nppes_provider_country','Is Sole Proprietor'])

### Create DG Nodes and Provider Type Relationships

# create DG nodes
df = df.unique()
df.rename({'npi':'providerID:ID'})
df[':LABEL'] = 'Provider'
df.save('data/graph/providers.csv', format='csv')

#create provider type relationships
df3 = docgraph_node_raw.select_columns(['npi','provider_type'])
df3.rename({'npi':':START_ID','provider_type':':END_ID'})
df3[':TYPE']='Is_Provider_Type'
df3[':END_ID'] = df3[':END_ID'].apply(lambda x: "spec_" + x.lower().lstrip().rstrip().replace(' ','_'))
df3 = df3.unique()
df3.save('data/graph/pt_rels.csv', format='csv')

## Create HCPCS Nodes and Rels

# import hcpcs data created previously
hcpcs_desc = gl.SFrame.read_csv('data/hcpcs_descriptions.csv',verbose=False)

# create hcpcs nodes
df1 = docgraph_node_raw[['hcpcs_code']]
df1.num_rows()
df1 = df1.join(hcpcs_desc,'hcpcs_code')
df1.rename({'hcpcs_code':'hcpcs_codeID:ID'})
df1[':LABEL']='HCPCS_Code'
df1 = df1.unique()
df1.save('data/graph/hcpcs_codes.csv', format='csv')

#create hcpcs relationships
df2 = docgraph_node_raw.select_columns(['npi','place_of_Service','line_srvc_cnt','bene_unique_cnt','bene_day_srvc_cnt',
                                        'average_Medicare_allowed_amt','stdev_Medicare_allowed_amt',
                                        'average_submitted_chrg_amt','stdev_submitted_chrg_amt',
                                        'average_Medicare_payment_amt','stdev_Medicare_payment_amt','hcpcs_code'])
df2 = df2.unique()
df2.rename({'npi':':START_ID','hcpcs_code':':END_ID'})
df2[':TYPE']='Conducts_Procedure'
df2.save('data/graph/hcpcs_rels.csv', format='csv')

## Create Taxonomy Nodes and Rels

#create taxonomy code nodes
tax_info = gl.SFrame.read_csv('data/nucc_taxonomy_150.csv',verbose=False)
tax_info = tax_info.remove_columns(['Definition','Notes']).rename({'Code':'taxonomy_codeID:ID'})
tax_info[':LABEL'] = 'Taxonomy_Code'
tax_info.save('data/graph/taxonomies.csv')

#create taxonomy relationships
df3 = docgraph_node_raw.select_columns(['npi','Healthcare Provider Taxonomy Code_1'])
df3.rename({'npi':':START_ID','Healthcare Provider Taxonomy Code_1':':END_ID'})
df3[':TYPE']='Has_Taxonomy'
df3 = df3.unique()
df3.save('data/graph/taxonomy_rels.csv', format='csv')

## Create Referral Relationships between Providers

#create referral relationships
edges = gl.SFrame.read_csv('data/dg_edges_chicago.csv',delimiter=',',
                           column_type_hints={'SharedTransactionCount':int,'PatientTotal':int, 'SameDayTotal':int},
                           verbose=False)

edges = edges.select_columns(['FirstNPI','SecondNPI','SharedTransactionCount'])
edges.rename({'FirstNPI':':START_ID','SecondNPI':':END_ID'})
edges[':TYPE']='Refers_To'

provider_dict = {}
for npi in df['providerID:ID']:
    provider_dict[npi] = 0

npi_list = {}
for npi in edges[':START_ID'].unique().append(edges[':END_ID'].unique()).unique():
    npi_list[npi] = 0

new_edges = edges.filter_by(npi_list.keys(),':START_ID').filter_by(npi_list.keys(),':END_ID')
new_edges.save('data/graph/provider_rels.csv', format='csv')

# Create SKA Nodes and Rels

- Affiliation Value / State
- IPA Value, State
- Hospital Value, State
- Health System Value, State
- PGP Value, State
- Physician Specialty Value
- Practice Specialty Value
- DEPT_EXPL

### Import the SK&A Data

# import data, change NPI to a string, filter the data by NPIs in Provider nodes
ska_data = gl.SFrame.read_csv('data/SK&A Data Decoded.csv',verbose=False)
ska_data['npi'] = ska_data['npi'].astype(str)
ska_data = ska_data.filter_by(npi_list.keys(),'npi',exclude=True)

# rename the data
ska_data = ska_data.rename({'T1':'Medical Title','COMPANY1':'Company','SIZE':'Num of Docs','PATVOLN':'Avg Daily Patient Volume',
                 'FS':'Free Standing Surgery'})

### Create SKA Nodes

# create Medical Title nodes
med_title = ska_data[['Medical Title']]
med_title.rename({'Medical Title':'medical_title'})
med_title[':LABEL'] = 'Medical_Title'
med_title['medical_titleID:ID'] = med_title['medical_title'].apply(lambda x: "mt_" + x.lower().lstrip().rstrip().replace(' ','_'))
med_title = med_title.unique()
med_title.save('data/graph/medical_titles.csv')

# create specialty nodes (covers provider type and physician/practice specialty)
phys_spec = ska_data.select_column('Physician Specialty Value')
prac_spec = ska_data.select_column('Practice Specialty Value')
p_t = docgraph_node_raw.select_column('provider_type')
spec_nodes = phys_spec.append(prac_spec).append(p_t)
spec_nodes_df = gl.SFrame(spec_nodes.unique()).rename({'X1':'specialty'})
spec_nodes_df[':LABEL'] = 'Specialty'
spec_nodes_df['specialtyID:ID'] = spec_nodes_df['specialty'].apply(lambda x: "spec_" + x.lower().lstrip().rstrip().replace(' ','_'))
spec_nodes_df = spec_nodes_df.unique()
spec_nodes_df.save('data/graph/specialties.csv')

# create Entity nodes
aff = ska_data[['Affiliation Value']]
aff['Origination'] = 'Affiliation'
aff.rename({'Affiliation Value':'entity_name'})
aff['entityID:ID'] = aff['entity_name'].apply(lambda x: "ent_" + x.lower().lstrip().rstrip().replace(' ','_'))
aff = aff.unique()

ipa = ska_data[['IPA Value']]
ipa['Origination'] = 'IPA'
ipa.rename({'IPA Value':'entity_name'})
ipa['entityID:ID'] = ipa['entity_name'].apply(lambda x: "ent_" + x.lower().lstrip().rstrip().replace(' ','_'))
ipa = ipa.unique()

hosp = ska_data[['Hospital Value']]
hosp['Origination'] = 'Hospital'
hosp.rename({'Hospital Value':'entity_name'})
hosp['entityID:ID'] = hosp['entity_name'].apply(lambda x: "ent_" + x.lower().lstrip().rstrip().replace(' ','_'))
hosp = hosp.unique()

hs = ska_data[['Health System Value']]
hs['Origination'] = 'Health System'
hs.rename({'Health System Value':'entity_name'})
hs['entityID:ID'] = hs['entity_name'].apply(lambda x: "ent_" + x.lower().lstrip().rstrip().replace(' ','_'))
hs = hs.unique()

pgp = ska_data[['PGP Value']]
pgp['Origination'] = 'PGP'
pgp.rename({'PGP Value':'entity_name'})
pgp['entityID:ID'] = pgp['entity_name'].apply(lambda x: "ent_" + x.lower().lstrip().rstrip().replace(' ','_'))
pgp = pgp.unique()

comp = ska_data[['Company']]
comp['Origination'] = 'Company'
comp.rename({'Company':'entity_name'})
comp['entityID:ID'] = comp['entity_name'].apply(lambda x: "ent_" + x.lower().lstrip().rstrip().replace(' ','_'))
comp = comp.unique()

entities = aff.append(ipa).append(hosp).append(hs).append(pgp).append(comp)
entities = entities[['entityID:ID','entity_name']]
entities[':LABEL'] = 'Entity'
entities['entity_name'] = entities['entity_name'].apply(lambda x: x.lower())
entities = entities.unique()
entities.save('data/graph/entities.csv')

# create Career Title nodes
df = ska_data.select_columns(['DEPT_EXPL','DEPTCODE'])
df.rename({'DEPT_EXPL':'career_title','DEPTCODE':'deptID:ID'})
df[':LABEL'] = 'Career_Title'
df['career_title'] = df['career_title'].apply(lambda x: 'House-Call Doctor' if x == 'House-call Doctors' else x)
df = df.unique()
df.save('data/graph/career_titles.csv', format='csv')

### Create SKA Relationships

# create affiliation rels
df = ska_data.select_columns(['npi','Affiliation Value','Affiliation State'])
df.rename({'npi':':START_ID','Affiliation Value':':END_ID','Affiliation State':'State'})
df[':TYPE']='Hospital_Affiliation'
df[':END_ID'] = df[':END_ID'].apply(lambda x: "ent_" + x.lower().lstrip().rstrip().replace(' ','_'))
df = df.unique()
df.save('data/graph/affiliation_rels.csv', format='csv')

# create IPA rels
df = ska_data.select_columns(['npi','IPA Value','IPA State'])
df.rename({'npi':':START_ID','IPA Value':':END_ID','IPA State':'State'})
df[':TYPE']='Part_of_IPA'
df[':END_ID'] = df[':END_ID'].apply(lambda x: "ent_" + x.lower().lstrip().rstrip().replace(' ','_'))
df = df.unique()
df.save('data/graph/IPA_rels.csv', format='csv')

# create Hospital rels
df = ska_data.select_columns(['npi','Hospital Value','Hospital State'])
df.rename({'npi':':START_ID','Hospital Value':':END_ID','Hospital State':'State'})
df[':TYPE']='Hospital_Ownership'
df[':END_ID'] = df[':END_ID'].apply(lambda x: "ent_" + x.lower().lstrip().rstrip().replace(' ','_'))
df = df.unique()
df.save('data/graph/hospital_rels.csv', format='csv')

# create Health System rels
df = ska_data.select_columns(['npi','Health System Value','Health System State'])
df.rename({'npi':':START_ID','Health System Value':':END_ID','Health System State':'State'})
df[':TYPE']='Health_System_Used'
df[':END_ID'] = df[':END_ID'].apply(lambda x: "ent_" + x.lower().lstrip().rstrip().replace(' ','_'))
df = df.unique()
df.save('data/graph/health_system_rels.csv', format='csv')

# create PGP rels
df = ska_data.select_columns(['npi','PGP Value','PGP State','CD3_SITE','CD3ALL_DRS'])
df.rename({'npi':':START_ID','PGP Value':':END_ID','PGP State':'State','CD3_SITE':'Num_Sites','CD3ALL_DRS':'Num_Docs'})
df[':TYPE']='PGP_Affiliation'
df[':END_ID'] = df[':END_ID'].apply(lambda x: "ent_" + x.lower().lstrip().rstrip().replace(' ','_'))
df = df.unique()
df.save('data/graph/pgp_rels.csv', format='csv')

# create company rels
df = ska_data.select_columns(['npi','Company','Free Standing Surgery','Num of Docs','Avg Daily Patient Volume',
                              'LATITUDE','LONGITUDE','ADDRESS1','CITY','STATE','ZIP'])
df.rename({'npi':':START_ID','Company':':END_ID','LATITUDE':'lat','LONGITUDE':'long',
           'ADDRESS1':'Address','CITY':'City','STATE':'State','ZIP':'Zip'})
df[':TYPE']='Part_of_Company'
df[':END_ID'] = df[':END_ID'].apply(lambda x: "ent_" + x.lower().lstrip().rstrip().replace(' ','_'))
df = df.unique()
df.save('data/graph/company_rels.csv')

# create Physician Specialty rels
df = ska_data.select_columns(['npi','Physician Specialty Value'])
df.rename({'npi':':START_ID','Physician Specialty Value':':END_ID'})
df[':TYPE']='Has_Physician_Specialty'
df = df.unique()
df[':END_ID'] = df[':END_ID'].apply(lambda x: "spec_" + x.lower().lstrip().rstrip().replace(' ','_'))
df = df.unique()
df.save('data/graph/phys_spec_rels.csv', format='csv')

# create Practice Specialty rels
df = ska_data.select_columns(['npi','Practice Specialty Value'])
df.rename({'npi':':START_ID','Practice Specialty Value':':END_ID'})
df[':TYPE']='Has_Practice_Specialty'
df = df.unique()
df[':END_ID'] = df[':END_ID'].apply(lambda x: "spec_" + x.lower().lstrip().rstrip().replace(' ','_'))
df = df.unique()
df.save('data/graph/prac_spec_rels.csv', format='csv')

# create Career Title rels
df = ska_data.select_columns(['npi','DEPTCODE'])
df.rename({'npi':':START_ID','DEPTCODE':':END_ID'})
df[':TYPE']='Has_Career_Title'
df = df.unique()
df.save('data/graph/career_rels.csv', format='csv')

# create Medical Title rels
df = ska_data.select_columns(['npi','Medical Title'])
df.rename({'npi':':START_ID','Medical Title':':END_ID'})
df[':TYPE']='Has_Medical_Title'
df[':END_ID'] = df[':END_ID'].apply(lambda x: "mt_" + x.lower().lstrip().rstrip().replace(' ','_'))
df = df.unique()
df.save('data/graph/medical_rels.csv', format='csv')

<p>bin/neo4j-import --into graph.db/ --stacktrace --nodes taxonomies.csv --relationships taxonomy_rels.csv --nodes specialties.csv --relationships pt_rels.csv --relationships prac_spec_rels.csv --relationships phys_spec_rels.csv --nodes providers.csv --relationships provider_rels.csv --nodes medical_titles.csv --relationships medical_rels.csv --nodes entities.csv --relationships pgp_rels.csv --relationships IPA_rels.csv --relationships hospital_rels.csv --relationships health_system_rels.csv --relationships affiliation_rels.csv --relationships company_rels.csv --nodes hcpcs_codes.csv --relationships hcpcs_rels.csv --nodes career_titles.csv --relationships career_rels.csv
