==> doc graph/dataset_builder/6_neo4j graph.db Creation.ipynb <==
# neo4j Graph Database Creation

After the execution of notebook #5, there should be multiple files in the _'graph'_ folder. Now we need to create the _'graph.db'_ folder to put into our neo4j database instance. See instructions below...

## Compiling Instructions
1. In your terminal, navigate to the _'data/graph/'_ directory
    - e.g. <code> $ cd data/graph/ </code>
2. In this directory, execute the following block of code:
    - <code><b>~/your/path/to/neo4j/</b>bin/neo4j-import --into graph.db/ --stacktrace --nodes taxonomies.csv --relationships taxonomy_rels.csv --nodes specialties.csv --relationships pt_rels.csv --relationships prac_spec_rels.csv --relationships phys_spec_rels.csv --nodes providers.csv --relationships provider_rels.csv --nodes medical_titles.csv --relationships medical_rels.csv --nodes entities.csv --relationships pgp_rels.csv --relationships IPA_rels.csv --relationships hospital_rels.csv --relationships health_system_rels.csv --relationships affiliation_rels.csv --relationships company_rels.csv --nodes hcpcs_codes.csv --relationships hcpcs_rels.csv --nodes career_titles.csv --relationships career_rels.csv</code>
    - Replace <code><b>~/your/path/to/neo4j/</b></code> above with the path to the neo4j instance installed on your machine
        - e.g. <code>~/Documents/Packages/neo4j-community-2.2.0/</code>
    - After you execute this code, it will take roughly 10-20 seconds to compile
3. Copy the _'graph.db'_ folder in this directory
4. Paste the _'graph.db'_ folder into your neo4j instance's directory, in the <code>/data/</code> folder
    - the final directory should be <code>~/your/path/to/neo4j/data/graph.db</code>
        - e.g. <code>~/Documents/Packages/neo4j-community-2.2.0/data/graph.db</code>
5. Start your neo4j database
6. Go to a web browser and navigate to localhost:7474
